Move Color utility class to libraries instead of constants
